refactor: gc to gradient_centralization
